fixed repaint issue for adding new tasks
