+++ client/src/utils/actions.js
@@ -1,7 +1,7 @@
 export const ADD_TIMESHEET_TASK = "ADD_TIMESHEET_TASK";
 export const CLEAR_TIMESHEET_TASK = "CLEAR_TIMESHEET_TASK";
 export const SHOW_ALERT_MODAL = "SHOW_ALERT_MODAL";
-export const CLEAR_ALERT_MODAL = "CELAR_ALERT_MODAL";
+export const CLEAR_ALERT_MODAL = "CLEAR_ALERT_MODAL";
 export const OPEN_ADD_EMPLOYEE_MODAL = "OPEN_ADD_EMPLOYEE_MODAL";
-export const FORCE_RENDER = "FORCE_RENDER";
+export const UPDATE_TODO_COUNT = "UPDATE_TODO_COUNT";
 export const UPDATE_DASHBOARD_TASKS = "UPDATE_DASHBOARD_TASKS";
