+++ client/src/utils/queries.js
@@ -78,6 +78,7 @@ export const QUERY_PROJECT = gql`
             tasks {
                 _id
                 title
+                description
                 employees {
                     _id
                     username
